accelerometer raw counts per unit was getting truncated- changed to return by float value. Unit tests save

+++ include/accelerometer/accelerometer_device.h
@@ -12,6 +12,6 @@ void accelerometer_device_init();
 
 unsigned int accelerometer_device_read(unsigned int channel);
 
-unsigned int accelerometer_device_counts_per_unit(unsigned int channel);
+float accelerometer_device_counts_per_unit(unsigned int channel);
 
 #endif /* ACCELEROMETER_DEVICE_H_ */
